[#29] fix some errors with licence=null
When a licence entry is null in mdl_files table, the oer plugin stopped
working. As fallback the subplugins folder and resource now test for
null value and use licence 'unknown' for that cases.
The release also was dependent from that error, now the release class
does not need to load the elements anymore, only the metadata from the
snapshot table is relevant. The element loading has been moved to the
zipper, as only there it is needed. Should improve performance of
public_metadata service a lot.
A bug was fixed in upgrade script, when files do not exist anymore but
are linked in oer plugin.